feat(thirdparty): adjust callback endpoint logic
The callback endpoint now appends a one time token to the redirect URL
instead of setting the JWT (in cookie or in header) directly. The
transaction boundary had to be extended to account for token generation
and now includes the entirety of operations in the callback,
including the linkAccount function, which now receives a connection as
an argument.