refactor(cmd): rework config initalization
Config initialization is done in the root command but some commands
do not require a config. If these commands are executed locally in the
checked out repo they work fine, but they will fail if they are
executed e.g. using the Docker image. I would have to provide
a config file path to the container and mount the referenced
file into the container even though it is not needed for the command.

This commit changes this by making those commands that require a
config load it themselves. It also pulls down the config flag to
the individual commands that require it instead of using a persistent
flag on the root level.