ci: remove docs auto deploy on backend tag
The gh-pages action as it is configured seems to reset the custom domain value
to docs.hanko.io in the repo's pages settings. This somehow leads to conflicts with
the new docs because the static JS docs must be reachable at a teamhanko.github.io
URL and not docs.hanko.io.