fix: set cookie secure flag
Safari does not allow setting secure cookies on localhost.
The secure flag is currently set based on the protocol of the
API only. With these changes, the protocol of the current
browser location is also checked, i.e. only if both API and
current location use HTTPS the cookie has the secure flag set
to true.