fix: skip on passkey onboarding in registration flow
The hook creating users is not aware of whether a user skipped
passkey onboarding. In case of a skip no credential is created
but the hook assumes there is a stringified credential avalable
on the stash. It attempts to unmarshal it but this fails.

This commit fixes it by first checking if a credential exists
on the stash.