feat: rework flow api profile action initializations
When generating the response after an action execution, schema generation and
initialization of actions for the next state are skipped if the execution result
already contains an execution schema (which should be the case if the action
executed successfully). This leads to problems when the action that was executed
is also an action for the next state (which is the case for profile actions,
where e.g. after deleting an email the flow returns to the profile init state,
which again may offer to execute the same action again, i.e. delete additional
emails): if the action execution leads to a state (of data) where the "next
execution" for the same action then should be suspended and the action should not
be part of the response (e.g. when trying to delete the last webauthn credential
if webauthn is the only auth method configured), then skipping the initialization
results in creating a response based on old/stale data, i.e. the action is part
of the response and is "offered" as a valid action when it actually should not.

This commit fixes that by removing the check for an existing exexution. This has
the drawback that the action is initialized twice.