feat: share email verification hook
Rename the hook to make it more obvious that it is resonsible for
persisting the verification status for an email after verification.

Move it to the shared package because the login flow also needs it,
e.g. if email verification was activated retroactively and users
logging in with previously unverified email addresses also correctly
have said email's verification status set to verified.

Add a noop case if the email address already exists and is also
already verfified.