fix: restore restore schema generation logic
Skipping schema generation for the next state actions if the schema
already existed was previously removed. This results in errors added
to inputs during execution not being part of the response because
a new schema without error information was generated. This change
reverts this.

If any action changes data such that it leads to a state of data
where the same action should be suspended for the next execution
(esp. in case of the profile where successful actions result in the
same state as the previous one and where the same action might be
part of the available actions) then the action can/should now check
the new state of data for suspension during execution and then set
a flag on the context/execution result accordingly (capabilites to
do so are introduced with these changes). Schema generation is then
skipped if there already is a schema and if the suspension flag is
set to true.