fix: do not make provider display name configurable
The display name is currently returned as part of the allowed
values for the inputs of the thirdparty oauth action. The value
of the input is the display name lowercased. If the display
name is configurable then it can be set to a value
for which there is no provider available. E.g. a user sets
the display name of the google provider to "MyCoolProvider".
Then the allowed value is "mycoolprovider". The provider
lookup will then fail, because lookup is based on the config
keys.