fix: registration success hook execution order
Hook execution order is the reverse of the actual argument order,
so the issue session hook is run first and user creation is run
last. This leads to the issue session hook not having the full
data for the created user and hence the JWT also does not contain
email of the user. This commit changes the order so that the
issue session hook is run last.