fix: fix security key login
The MFA login with security key due to wrongly used validation method of the webauthn library. The security key validation must not use the validation method for discoverable credentials, because the MFA credential is not created as a discoverable credential. Because of the usage of the wrong method a MFA login with a security key never succeeds.