feat: add custom claims to session validate endpoint responses
Add custom claims on top level of returned claims object (i.e.
instead of storing them under a dedicated key like "custom_claims").
Uses a custom marshaller for the claims struct that does this
"flattening". An alternative approach of unmarshalling the token into
a claims struct instance with custom claims using encoding/json's
"inline" struct tag does not work because the type of custom claims
is unknown and it is hence a map type but inlining seems to work
for structs with named fields only.