Fix login method chooser (#2111)
* fix: remove password option from method chooser

When Privacy.OnlyShowActualLoginMethods config setting is enabled and the user has no password or passwords are disabled the password option must not be available in the login method chooser.

* fix: remove passcode option from method chooser

When Privacy.OnlyShowActualLoginMethods config setting is enabled and the user has either no email address or emails are disabled or emails can not be used for authentication, the passcode option must not be available in the login method chooser.