fix: backend handling of prefilled username and email
Prefilling both username and email even though only one of them
is enabled in the backend can lead to unwanted effects on registration.

For example, if emails are enabled (and acquired) but usernames are
not, and a user prefills both email and username in elements, then
the backend action does not properly check the enabled status for
the identifier type which can lead to creation of a username or an
email even though the respective identifier type is not enabled.
It can also lead to misleading errors, e.g. returning an error because
the check for an existing username still runs and fails even though
usernames are disabled.

This commit adds apropriate checks on identifier types (enabled and
acquisition on registration) where necessary.